+++ primal-test/memory-tests/Test/Primal/Memory/BytesSpec.hs
@@ -18,23 +18,13 @@ import Control.DeepSeq
 import Control.Monad
 import Data.ByteString.Builder
 import qualified Data.ByteString.Lazy.Char8 as BSL8
-import Data.Complex
-import Data.Functor.Identity
 import qualified Data.List as List
-import Data.Ratio
 import Foreign.Storable
-import GHC.Fingerprint.Type
-import GHC.IO.Device
 import Numeric
-import Data.Semigroup
 import Primal.Array
 import Primal.Foreign hiding (Any)
-import Primal.Foreign.StablePtr
 import Primal.Memory.Addr
-import Primal.Memory.ByteString
 import Primal.Memory.Bytes
-import Primal.Memory.PArray
-import Primal.Memory.Text
 import System.Timeout
 import Test.Primal
 import Test.Primal.Memory